Move fetchServerResponse into a separate file
