+++ packages/next/src/client/components/router-reducer/fill-cache-with-data-property.ts
@@ -1,5 +1,5 @@
 import { CacheNode, CacheStates } from '../../../shared/lib/app-router-context'
-import { fetchServerResponse } from '../app-router'
+import { fetchServerResponse } from './fetch-server-response'
 
 /**
  * Kick off fetch based on the common layout between two routes. Fill cache with data property holding the in-progress fetch.
